Use import options
The import results vary depending on user preferences because the
`OpenInsertObject` does not support settings any options (it uses the
user preferences).
To ensure consistent results, we save the current preferences, set the
desired ones, import the model, and restore the previous preferences.
Add the `import_original` function to handle this.